+++ examples/vue-example/main.vue
@@ -8,7 +8,7 @@
         This will be shown after load.
       </v-loading>
     </div>
-    <button @click='writeCode()' :disable='$vueLoading.isLoading("writing code")'>
+    <button @click='writeCode()' v-loading:disabled='"writing code"'>
       <v-loading loader='writing code' message='Writing Code...'>
         <template slot='loading'>
           <span class="animated">️️♻️</span>
@@ -16,22 +16,22 @@
         Write Code
       </v-loading>
     </button>
-    <button @click='endLoading()' :disabled='!$vueLoading.isLoading("writing*")'>
-      <span v-if='$vueLoading.isLoading("writing*")'>Stop Coding</span>
-      <span v-else>Coding Stopped</span>
+    <button v-loading:click.end='"writing code"'>
+      <span v-loading:visible="'writing *'">Stop Coding</span>
+      <span v-loading:visible.not="'writing *'">Coding Stopped</span>
     </button>
     <br>
-    <button @click='$vueLoading.startLoading("c")' :disabled='$vueLoading.isLoading("c")'>
+    <button v-loading:click.start='"c"' v-loading:disabled='"c"'>
       Start <b>"c"</b> Loader
     </button>
-    <button @click='$vueLoading.endLoading("c")' :disabled='!$vueLoading.isLoading("c")'>
+    <button v-loading:click.end='"c"' v-loading:disabled.not='"c"'>
       Stop <b>"c"</b> Loader
     </button>
     <p>
       Toggle on any letter to load/unload together.
     </p>
     <ul class="list">
-      <li v-for='loader in loaders' :key='loader + Math.random()' @click='toggleLoader(loader)'>
+      <li v-for='loader in loaders' :key='loader + Math.random()' v-loading:toggle='loader'>
         <v-loading :loader='loader'>
           <template slot='loading' v-if='loader == "c"'>
             <span class="animated">🎉</span>
@@ -60,13 +60,6 @@
       },
       endLoading() {
         this.$vueLoading.endLoading('writing code');
-      },
-      toggleLoader(loader) {
-        if (this.$vueLoading.isLoading(loader)) {
-          this.$vueLoading.endLoading(loader);
-        } else {
-          this.$vueLoading.startLoading(loader);
-        }
       }
     }
   }